--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ tinysl
 cover.out
 cover.cov
 coverprofile.out
+*.test