… .github/workflows/codeql-analysis.yml
… .github/workflows/go.yml
… .github/workflows/golang-test-annotations.yml
… README.md
… constructor.go
… container.go
… container_test.go
… doc.go
… errors.go
… go.mod
… go.sum
… locator.go
… locator_test.go
… scope.go
… tinsl_bench_test.go
… tinsl_total_bench_test.go
… tinysl.go
… tinysl_bench_test.go
… tinysl_test.go
… tinysl_total_bench_test.go
… utils_test.go