## New Features:

- New types of constructors:
  * returning just type without error: `T`
  * returning `Cleanup` as second argument in result: `(T, Cleanup, error)`
- `Logger` to handle panics during cleanup
- `DecorateMiddleware` to decorate go http middleware
- method `Replace` for `Container` to replace already registered implementations 
- method `Decorate` for `Container` to decorate already registered implementations 

## Updates and changes:

- Changes `Lifetime` underlying type from `string` to `int`, but adds `Stringer` implementation
- Changes `DecorateHandler` to return `http.HandlerFunc` instead of `http.Handler` and accept any implementation of `http.Handler`
- Changes the way `PerContext` services references are cleared
- Adds cleanup workers for singleton and per-context scopes
- Catches panic on constructor call for new service and converts to error
- Catches panic in cleanup functions and logs the reason

## Cleanup workers overview:

Worker is spawned as separate goroutine.
There is one worker for singletons and number of workers for per-context. Numbers of workers for per-context is configured using `WithWorkerPoolSize` `ContainerOption`. Bu default it is set to the number of CPUs.
When context is done cleanup of the bounded services is done in those workers. All custom cleanup functions returned by services constructors are also handled in those workers after the context is done.
For singleton services cleanup is called when `os.Interrupt`, `syscall.SIGTERM` or `syscall.SIGINT` is called. 
Alternatively user can provide underlying context using `WithSingletonCleanupContext` `ContainerOption` when creating new `Container`. When this context is cancelled - cleanup functions for singleton services and remaining uncleaned per-context services will be called.

## Motivation for changes in instance lookup logic (`concurrecny.go`)

On large amount of parallel requests memory usage is really close, but time spent on one iteration is significantly better with new implementation:

Command: `go test -benchmem -bench=. ./... -benchtime=10s`

### Old Implementation with old lookup:

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetTransinet-8                                 50034876               207.0 ns/op            80 B/op          3 allocs/op
BenchmarkGetPerContext-8                                 1582528              7655 ns/op             565 B/op         16 allocs/op
BenchmarkGetPerContext2Services-8                        1000000             15502 ns/op            1003 B/op         30 allocs/op
BenchmarkGetPerContext2Contexts-8                        1000000             11584 ns/op            1122 B/op         32 allocs/op
BenchmarkGetPerContext2Services2Contexts-8                669417             18743 ns/op            2023 B/op         60 allocs/op
BenchmarkGetPerContext2Services10Contexts-8               668148             18836 ns/op            2026 B/op         60 allocs/op
BenchmarkGetPerContext4Services10Contexts-8                80586            151829 ns/op           17201 B/op        519 allocs/op
BenchmarkGetPerContext4Services100Contexts-8                7069           1795268 ns/op          173750 B/op       5189 allocs/op
BenchmarkGetPerContext4Services1000Contexts-8                172          68369982 ns/op         1764215 B/op      52059 allocs/op
BenchmarkGetPerContext4Services10_000Contexts-8                2        6187663062 ns/op        22525412 B/op     544871 allocs/op
BenchmarkGetSingleton-8                                 277002632               43.32 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       164.244s
```

### New implementation with old lookup:

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetTransinet-8                                 50053624               260.2 ns/op            80 B/op          3 allocs/op
BenchmarkGetPerContext-8                                 1935006              6085 ns/op             848 B/op         27 allocs/op
BenchmarkGetPerContext2Services-8                         953317             13850 ns/op            1529 B/op         51 allocs/op
BenchmarkGetPerContext2Contexts-8                        1000000             11499 ns/op            1660 B/op         52 allocs/op
BenchmarkGetPerContext2Services2Contexts-8                511026             23948 ns/op            3035 B/op        101 allocs/op
BenchmarkGetPerContext2Services10Contexts-8               486147             25053 ns/op            3043 B/op        101 allocs/op
BenchmarkGetPerContext4Services10Contexts-8                68716            173992 ns/op           28068 B/op        804 allocs/op
BenchmarkGetPerContext4Services100Contexts-8                1216         134929475 ns/op          575312 B/op       7043 allocs/op
BenchmarkGetPerContext4Services1000Contexts-8                100         234465171 ns/op         6510368 B/op      69285 allocs/op
BenchmarkGetPerContext4Services10_000Contexts-8               50        2474863602 ns/op        63312934 B/op     694488 allocs/op
BenchmarkGetSingleton-8                                 232203060               52.88 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       489.592s
```

### New implementation with new lookup:

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetTransinet-8                                 49611889               241.2 ns/op            80 B/op          3 allocs/op
BenchmarkGetPerContext-8                                 2061027              5350 ns/op             765 B/op         25 allocs/op
BenchmarkGetPerContext2Services-8                         932289             11744 ns/op            1436 B/op         48 allocs/op
BenchmarkGetPerContext2Contexts-8                        1000000             11968 ns/op            1518 B/op         50 allocs/op
BenchmarkGetPerContext2Services2Contexts-8                482998             24225 ns/op            2856 B/op         95 allocs/op
BenchmarkGetPerContext2Services10Contexts-8               491896             24700 ns/op            2847 B/op         95 allocs/op
BenchmarkGetPerContext4Services10Contexts-8                69286            172607 ns/op           26258 B/op        745 allocs/op
BenchmarkGetPerContext4Services100Contexts-8               10000           1459015 ns/op          410407 B/op       6745 allocs/op
BenchmarkGetPerContext4Services1000Contexts-8                850          15344904 ns/op         3194739 B/op      60549 allocs/op
BenchmarkGetPerContext4Services10_000Contexts-8               66         169801989 ns/op        26553210 B/op     587404 allocs/op
BenchmarkGetSingleton-8                                 244166288               48.36 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       255.632s
```

Separately here is bench test for `100_000` parallel requests to get `PerContext` service:

### New implementation with new lookup (100 000 Contexts):

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetPerContext4Services100_000Contexts-8               8        1476122021 ns/op        481730466 B/op   6452758 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       32.438s
```

## Update

Moving to `sync.Map` in managing access to instances brought event more performance (in benchmarks at least):

### New implementation with new lookup and new instance access:

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetTransinet-8                                 54590997               224.3 ns/op            80 B/op          3 allocs/op
BenchmarkGetPerContext-8                                 2101644              5612 ns/op             943 B/op         32 allocs/op
BenchmarkGetPerContext2Services-8                        1223041              9855 ns/op            1453 B/op         50 allocs/op
BenchmarkGetPerContext2Contexts-8                         984096             12023 ns/op            1864 B/op         65 allocs/op
BenchmarkGetPerContext2Services2Contexts-8                639618             19879 ns/op            2968 B/op        101 allocs/op
BenchmarkGetPerContext2Services10Contexts-8               599737             20665 ns/op            2996 B/op        101 allocs/op
BenchmarkGetPerContext4Services10Contexts-8                97635            130858 ns/op           23314 B/op        732 allocs/op
BenchmarkGetPerContext4Services100Contexts-8               10000           1122171 ns/op          235561 B/op       7008 allocs/op
BenchmarkGetPerContext4Services1000Contexts-8               1287          15583586 ns/op         3412318 B/op      68411 allocs/op
BenchmarkGetPerContext4Services10_000Contexts-8               90         143006159 ns/op        27454336 B/op     652657 allocs/op
BenchmarkGetSingleton-8                                 126054327               90.36 ns/op           24 B/op          2 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       225.043s
```

### New implementation with new lookup  and new instance access (100 000 Contexts):

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetPerContext4Services100_000Contexts-8               8        1276476771 ns/op        267513037 B/op   6726685 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       33.233s
```

## More Updates:

Moving to using mostly pointers and reducing amount of string copying and comparisons proved to be another significant improvement both in memory consumption and speed of execution according to benchmark tests:

### Latest implementation:

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetTransinet-8                                 45034612               257.1 ns/op           104 B/op          4 allocs/op
BenchmarkGetPerContext-8                                 3090699              3888 ns/op             594 B/op         13 allocs/op
BenchmarkGetPerContext2Services-8                        2820310              3835 ns/op             634 B/op         15 allocs/op
BenchmarkGetPerContext2Contexts-8                        1729623              7288 ns/op            1202 B/op         28 allocs/op
BenchmarkGetPerContext2Services2Contexts-8               1541464              6554 ns/op            1234 B/op         29 allocs/op
BenchmarkGetPerContext2Services10Contexts-8              1609842              6697 ns/op            1251 B/op         30 allocs/op
BenchmarkGetPerContext4Services10Contexts-8               558772             18508 ns/op            5712 B/op        145 allocs/op
BenchmarkGetPerContext4Services100Contexts-8               87378            129067 ns/op           54568 B/op       1379 allocs/op
BenchmarkGetPerContext4Services1000Contexts-8               6255           1888807 ns/op          809482 B/op      21184 allocs/op
BenchmarkGetPerContext4Services10_000Contexts-8              765          15350472 ns/op         6225352 B/op     158523 allocs/op
BenchmarkGetSingleton-8                                 193900538               60.46 ns/op            8 B/op          1 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       212.297s
```

### Latest implementation (100 000 Contexts):

```
goos: darwin
goarch: arm64
pkg: github.com/andriiyaremenko/tinysl
cpu: Apple M1
BenchmarkGetPerContext4Services100_000Contexts-8              81         157843153 ns/op        64924666 B/op    1661427 allocs/op
PASS
ok      github.com/andriiyaremenko/tinysl       35.391s
```